requirements.txt
beautifulsoup4
requests
camelot-py[cv]
pandas
openpyxl
selenium